+++ extensions/slack/src/search-messages.tsx
@@ -1,30 +1,36 @@
 import { Action, ActionPanel, Color, Icon, List } from "@raycast/api";
 import { useCachedPromise, useCachedState } from "@raycast/utils";
-import { format } from "date-fns";
+import { format, formatDistanceToNow } from "date-fns";
 import { useState } from "react";
 import * as emoji from "node-emoji";
 
 import { withSlackClient } from "./shared/withSlackClient";
 import { getSlackWebClient } from "./shared/client/WebClient";
 import { convertTimestampToDate, handleError } from "./shared/utils";
-import { useChannels, useUsers } from "./shared/client";
+import { useChannels, useMe } from "./shared/client";
 import { SearchMessagesArguments } from "@slack/web-api";
 
 function Search() {
   const [query, setQuery] = useState("");
   const [selectedChannel, setSelectedChannel] = useState("");
+
   const [sortType, setSortType] = useCachedState<SearchMessagesArguments["sort"]>("search-sort-type", "timestamp");
 
-  const { data: users } = useUsers();
+  const { data: me } = useMe();
   const { data: channels } = useChannels();
+  const users = channels?.[0];
+
+  const meInfo = users?.find((u) => u.id === me?.id);
 
   const { data, isLoading } = useCachedPromise(
     async (query, channel, sort) => {
       const webClient = getSlackWebClient();
+
       const results = await webClient.search.messages({
-        query: `${query}${selectedChannel ? ` in:${channel}` : ""}`,
+        query: `${query}${channel ? ` in:${channel}` : ""}`,
         sort,
       });
+
       return results.messages?.matches ?? [];
     },
     [query, selectedChannel, sortType],
@@ -42,13 +48,26 @@ function Search() {
     { title: "Most Relevant", sort: "score" },
   ];
 
-  const FromMeOnlyAction = () => (
-    <Action
-      title="From Me Only"
-      icon={Icon.Person}
-      shortcut={{ modifiers: ["cmd", "shift"], key: "i" }}
-      onAction={() => setQuery(query ? `from:me ${query}` : "from:me ")}
-    />
+  const FromActions = () => (
+    <>
+      <Action
+        title="From Me Only"
+        icon={meInfo ? meInfo.icon : Icon.Person}
+        shortcut={{ modifiers: ["cmd", "shift"], key: "i" }}
+        onAction={() => setQuery(query ? `from:me ${query}` : "from:me ")}
+      />
+
+      <ActionPanel.Submenu icon={Icon.Person} title="From" shortcut={{ modifiers: ["cmd", "shift"], key: "u" }}>
+        {users?.map((u) => (
+          <Action
+            key={u.id}
+            title={u.name}
+            icon={u.icon}
+            onAction={() => setQuery(query ? `from:${u.name} ${query}` : `from:${u.name} `)}
+          />
+        ))}
+      </ActionPanel.Submenu>
+    </>
   );
 
   return (
@@ -58,38 +77,58 @@ function Search() {
       onSearchTextChange={setQuery}
       actions={
         <ActionPanel>
-          <FromMeOnlyAction />
+          <FromActions />
         </ActionPanel>
       }
       throttle
-      isShowingDetail
+      isShowingDetail={data && data.length > 0}
       searchBarAccessory={
-        <List.Dropdown tooltip="Search Channels" onChange={setSelectedChannel} storeValue>
+        <List.Dropdown tooltip="Search Channels" onChange={setSelectedChannel}>
           <List.Dropdown.Item value="" title="All Channels" />
-
-          <List.Dropdown.Section>
-            {channels?.flat().map((c) => {
-              return <List.Dropdown.Item key={c.id} icon={c.icon} value={c.name} title={c.name} />;
-            })}
-          </List.Dropdown.Section>
+          {channels ? (
+            <List.Dropdown.Section>
+              {channels.flat().map((c) => {
+                return <List.Dropdown.Item key={c.id} icon={c.icon} value={c.name} title={c.name} />;
+              })}
+            </List.Dropdown.Section>
+          ) : null}
         </List.Dropdown>
       }
     >
-      <List.EmptyView title={query.length > 0 && !isLoading ? "No messages found" : "Search for Slack Messages"} />
       {data?.map((m) => {
         if (!m.text || !m.ts) return null;
 
         const user = users?.find((u) => u.id === m.user);
         const date = convertTimestampToDate(m.ts);
         const text = emoji.emojify(m.text);
+        const formattedDate = format(date, "EEEE dd MMMM yyyy 'at' HH:mm");
 
         return (
           <List.Item
             key={m.iid}
             icon={{ value: user?.icon, tooltip: user?.name ?? "Unknown user" }}
             title={text}
-            accessories={[{ date, tooltip: format(date, "EEEE dd MMMM yyyy 'at' HH:mm") }]}
-            detail={<List.Item.Detail markdown={text} />}
+            accessories={[{ date, tooltip: formattedDate }]}
+            detail={
+              <List.Item.Detail
+                markdown={text}
+                metadata={
+                  <List.Item.Detail.Metadata>
+                    {m.channel ? (
+                      <List.Item.Detail.Metadata.Label title="Channel Name" icon={Icon.Hashtag} text={m.channel.name} />
+                    ) : null}
+
+                    {user ? <List.Item.Detail.Metadata.Label title="From" icon={user.icon} text={user.name} /> : null}
+
+                    <List.Item.Detail.Metadata.Label
+                      title="Posted"
+                      icon={Icon.Clock}
+                      text={formatDistanceToNow(date)}
+                    />
+                  </List.Item.Detail.Metadata>
+                }
+              />
+            }
             actions={
               <ActionPanel>
                 {m.permalink ? <Action.OpenInBrowser url={m.permalink} title="Open Message" /> : null}
@@ -111,7 +150,7 @@ function Search() {
                     ))}
                   </ActionPanel.Submenu>
 
-                  <FromMeOnlyAction />
+                  <FromActions />
                 </ActionPanel.Section>
               </ActionPanel>
             }