+++ extensions/slack/src/shared/client/SlackClientHooks.ts
@@ -2,13 +2,6 @@ import { handleError } from "../utils";
 import { SlackClient } from "./SlackClient";
 import { useCachedPromise } from "@raycast/utils";
 
-export const useUsers = () =>
-  useCachedPromise(SlackClient.getUsers, [], {
-    onError(error) {
-      handleError(error, "Failed to load users");
-    },
-  });
-
 export const useChannels = () =>
   useCachedPromise(
     () => Promise.all([SlackClient.getUsers(), SlackClient.getChannels(), SlackClient.getGroups()]),
@@ -19,6 +12,8 @@ export const useChannels = () =>
       },
     },
   );
+
+export const useMe = () => useCachedPromise(SlackClient.getMe);
 
 export const useUnreadConversations = (conversationIds: string[] | undefined) =>
   useCachedPromise((ids) => SlackClient.getUnreadConversations(ids), [conversationIds ?? []]);